feat: Implement caching for account and transaction retrieval, and add token utility functions
