+++ ahorratron/sync_api/main.py
@@ -1,27 +1,22 @@
-import datetime
-import json
 import logging
-import os
 from collections.abc import Awaitable, Callable
 
 import uvicorn
-from fastapi import Depends, FastAPI, Header, HTTPException, Request, Response
-from jose import jwe, jwt
+from fastapi import Depends, FastAPI, HTTPException, Request, Response
 
-from ahorratron.sync_api.connectors import get_connector
 from ahorratron.sync_api.models.account_models import Account, AccountsResponse
 from ahorratron.sync_api.models.core_models import SessionData, UserData
 from ahorratron.sync_api.models.transaction_models import TransactionsResponse
+from ahorratron.sync_api.service import (
+    get_account_by_id_cached,
+    get_accounts_cached,
+    get_transactions_cached,
+)
+from ahorratron.sync_api.utils.token import create_encrypted_token, get_decrypted_token
 
 logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
 
-SECRET_KEY = bytes.fromhex(os.environ["JWE_SECRET_KEY"])  # Must be 32 bytes for A256GCM
-TOKEN_DURATION = datetime.timedelta(hours=12)
-JWE_ALGORITHM = "A256GCM"
-
-JWT_SECRET_KEY = os.environ["JWT_SECRET_KEY"]
-JWT_ALGORITHM = "HS256"  # Or RS256 if you want asymmetric signing
 
 app = FastAPI()
 
@@ -42,43 +37,6 @@ async def log_request_middleware(
     return response
 
 
-# Generate encrypted JWT (JWE) from a dictionary
-def create_encrypted_token(data: SessionData) -> str:
-    enc_payload = data.model_dump()
-    enc_token = jwe.encrypt(
-        json.dumps(enc_payload), SECRET_KEY, algorithm=JWE_ALGORITHM
-    )
-    enc_token_str = enc_token.decode() if isinstance(enc_token, bytes) else enc_token
-
-    payload = {
-        "enc_token": enc_token_str,
-        "exp": int((datetime.datetime.now(datetime.UTC) + TOKEN_DURATION).timestamp()),
-    }
-    token = jwt.encode(payload, JWT_SECRET_KEY, algorithm=JWT_ALGORITHM)
-    return token
-
-
-# Dependency to get current user and password from encrypted token
-def get_decrypted_token(x_api_key: str = Header(..., alias="X-API-KEY")) -> SessionData:
-    try:
-        # Decode JWT
-        payload = jwt.decode(x_api_key, JWT_SECRET_KEY, algorithms=[JWT_ALGORITHM])
-        exp = payload.get("exp")
-        if exp and datetime.datetime.now(datetime.UTC).timestamp() > exp:
-            raise HTTPException(status_code=401, detail="Token expired")
-        enc_token = payload.get("enc_token")
-        if not enc_token:
-            raise HTTPException(status_code=401, detail="Missing encrypted token")
-        # Decrypt the encrypted token
-        decrypted = jwe.decrypt(enc_token, SECRET_KEY)
-        if decrypted is None:
-            raise HTTPException(status_code=401, detail="Invalid encrypted token")
-        session_data = SessionData.model_validate_json(decrypted.decode())
-        return session_data
-    except Exception:
-        raise HTTPException(status_code=401, detail="Invalid or expired token")
-
-
 @app.post("/auth")
 async def auth(request: UserData):
     data = SessionData(user_data=request)
@@ -93,8 +51,7 @@ async def auth(request: UserData):
 @app.get("/accounts", response_model=AccountsResponse)
 def get_accounts(itemId: str, session_data: SessionData = Depends(get_decrypted_token)):
     logger.debug(f"Session data: {session_data}")
-    connector = get_connector(session_data.user_data)
-    response = connector.get_accounts(itemId=itemId)
+    response = get_accounts_cached(session_data.user_data, itemId)
     logger.debug(f"Accounts response: {response}")
     return response
 
@@ -103,8 +60,7 @@ def get_accounts(itemId: str, session_data: SessionData = Depends(get_decrypted_
 def get_account_by_id(
     accountId: str, session_data: SessionData = Depends(get_decrypted_token)
 ):
-    connector = get_connector(session_data.user_data)
-    response = connector.get_account_by_id(accountId=accountId)
+    response = get_account_by_id_cached(session_data.user_data, accountId)
     logger.debug(f"Account detail response: {response}")
     return response
 
@@ -113,8 +69,7 @@ def get_account_by_id(
 def get_transactions(
     accountId: str, session_data: SessionData = Depends(get_decrypted_token)
 ):
-    connector = get_connector(session_data.user_data)
-    response = connector.get_transactions(accountId=accountId)
+    response = get_transactions_cached(session_data.user_data, accountId)
     logger.info(f"Transactions response: {response}")
     return response
